Prefer terraform-ls over terraform-lsp
The LSP terraform-ls is official from Hashicorp and is focused on
stability over experimental features, as terraform-lsp is focused. The
relationship between the two is described in the [terraform-ls README][]
and reiterated in the [terraform-lsp README][].

[terraform-ls README]: https://github.com/hashicorp/terraform-ls#terraform-ls-vs-terraform-lsp
[terraform-lsp README]: https://github.com/juliosueiras/terraform-lsp

Based on this, and the fact that both LSPs are well maintained,
I believe that vim-lsp-settings should prefer installing terraform-ls
over terraform-lsp.