Prevent `remark-language-server` for `markdown`
Resolves #543, #527

The current configuration for `remark-language-server` has it proposed
for installation whenever a markdown file is edited. This is incorrect,
as `remark-language-server` is _only_ for use in a `remark` project,
which implies a lot of extra baggage (node-base infrastructure,
configuration, etc.) not expected by most users.

As it does appears that `vim-lsp` and `vim-lsp-settings` allow only
filetype server identification by default, this PR:

- removes the `markdown` entry from the `allowlist` for
  `remark-language-server`;
- removes the `markdown` entry from `settings.json` (as
  `remark-language-server` was the only LS there, and
  `remark-language-server` is _not_ a markdown language server);
- and updates the compatibility listing so that `remark-language-server`
  is noted as `Markdown (remark)`.

My personal preference would be to wholly _remove_
`remark-language-server` (it provides value to a very small group of
developers), but as it _may_ be possible to extend the `allow` test in
the future so as to detect appropriate configuration, there is currently
no reason to remove it. This, however, depends on testing which I have
not yet completed related to whether `remark-language-server` can be
installed and started manually _without_ a filetype specification.