Add wrapper's configs for hie and hls
haskell-ide-engline(hie) and haskell-language-server(hls) are usually
called via their wrapper executables. However this plugin looks up
non-wrapper executables but not wrapper them in loading configurations,
so it doesn't load configs for hie and hls even if wrappers are
installed.

To resolve this, change command name which is used in loading configs
from non-wrappers to wrappers, and add configs for wrappers instead of
not-wrappers. And for compatibility, keep up loading user's non-wrapper
configs.