Change fsharp-language-server publish folder
Build and publish the fsharp-language-server to a path independent from
the .NET version.
Prior to this commit, the launch script assumes that the application
will be built to a folder whose path includes netcoreapp2.0. This means
that the script will fail if you build on a machine with, say, .NET 6.