fix: eclipse-jdt-ls member autocompletion work
Pressing a dot after typing an instance variable in a Java method body
to see its members and other definitions as suggestions is not working
currently in eclipse-jdt-ls for Java.

While this is working for imports, methods do not get resolved. This
commit replaces a suspicious dot that may be related to this, which
causes the issue to get fixed.